backend/requirements.txt
Flask
flask-cors
supabase
python-dotenv
psycopg2-binary
requests
pytest
gotrue
